[refactor] DesktopShellState をドメインスライス合成へ分割 (WP-H6 PR3)
- shell/slices/ に 11 ファイルを新設(shared + 10 ドメイン)。
  各スライスは XxxSliceState 型 + createInitialXxxSlice() を持ち、
  shared のみに依存する一方向構造(store.ts との循環なし)
- store.ts の DesktopShellState をスライスの交差型に、
  createInitialShellState() を createInitial*Slice() の spread 合成に置換
- 実行時は従来どおり単一 zustand store のまま。1 回の set() は
  スライスをまたいでも原子的で、useRouteSynchronization の
  一括更新 21 箇所は無変更
- 移動した型・定数・関数は store.ts から re-export し、
  既存 39 ファイルの import 面を維持(挙動変更なし)
- store.ts 552 行 → 178 行
